fix: On linux function rasters.unpackbits no longer over allocates memory

+++ sertit/rasters.py
@@ -854,7 +854,7 @@ def unpackbits(array: np.ndarray, nof_bits: int) -> np.ndarray:
     """
     xshape = list(array.shape)
     array = array.reshape([-1, 1])
-    msk = 2 ** np.arange(nof_bits).reshape([1, nof_bits])
+    msk = 2 ** np.arange(nof_bits, dtype=array.dtype).reshape([1, nof_bits])
     return (array & msk).astype(bool).astype(np.uint8).reshape(xshape + [nof_bits])
